Optimize some http code slightly
Three tweaks:
Assume files are large, so use a large buffer.
Simplify image download to use .net 4.5 api, stream.CopyTo, for copying
from the webresponse to the memory stream.
When testing if a url is valid, request just the HEAD, not the full Get.
That makes the round trip much quicker as the server doesn't start
sending data.